Fix broken namespace renaming in SIMD
